+++ spec/active_interaction/filters/record_filter_spec.rb
@@ -123,7 +123,7 @@ describe ActiveInteraction::RecordFilter, :filter do
       it 'raises an error' do
         expect do
           result
-        end.to raise_error ActiveInteraction::InvalidClassError
+        end.to raise_error ActiveInteraction::InvalidNameError
       end
     end